Manually merged Master and Graphing branches
Letting GitHub Desktop try to merge would result in a merge-error, so we did it manually.
First we copied the code I wrote in the Graphing branch and pasted that code in it's own seperate Processing files (so we could later copy  and check from there).
Then we pasted in what we could and manually added the small bits and pieces that couldn't be pasted in.
Then we tested to see if it still worked (which was scary and made me paranoid that I made a huge mistake somewhere - though I didn't)